M7 WI1+WI2(store): GC storage surfaces, epoch/gate protocol, sweep engine
Storage surfaces (plan 01): ShardedIndex repoint/remove/entries_for_pack
with pack guards; PageStore sealed_pack_ids/rotate_active/scan_pack/
read_record/create_gc_pack/delete_pack plus the shared next_pack_id
allocator (rotation refactored to use it — GC pack ids can no longer
collide with future rotations); GcPackWriter writes its sidecar from its
own record list, never via the index (empty-sidecar crash trap);
delete_pack unlinks .idx before .spk. Meta: version-stepping migration
runner (v1 DBs now actually step to v2), gc_state table, gc_root_refs /
list_tombstones / reap_tombstone (single-statement subtree delete — SQLite
immediate FKs check per statement) / set_gc_state, failpoints feature.

Store engine (plan 02): commit gate switched to parking_lot and widened
across parent-check→publish; gated validating register_live_ref (late
roots + full dependency-closure check, both put_snapshot exit paths);
begin/end_gc_epoch fence; gc.rs mark walk, pack sweep with
straggler-drain finalize under the sweep gate, batched manifest sweep
under the gate; GcHooks interleaving points + sabotage modes constructible
only under gc-test-hooks; failpoints gc-compact-copy/gc-index-repoint/
gc-pack-unlink/gc-manifest-unlink/gc-reap-txn.

Engine smoke tests: quiescent exactness (physical == reachable),
straggler commit at BeforeFinalize survives, R4 epoch latch. Workspace
fmt/clippy/tests green incl. failpoints matrix smoke (5 cycles).

+++ crates/snapstore-client/tests/page_channel_fallback.rs
@@ -197,7 +197,7 @@ async fn corrupt_cross_check_surfaces_mismatch() {
     let client = channel_client(&server).await;
 
     let pages: Vec<(u64, Vec<u8>)> = (0..4)
-        .map(|i| (i as u64, rand_page(0xBAD_C0DE, i)))
+        .map(|i| (i, rand_page(0xBAD_C0DE, i)))
         .collect();
 
     let result = client.put_pages(pages).await;
